Setup Godot to support the Oculus Mobile SDK.
