+++ modules/gdnative/gdnative_api.json
@@ -6286,7 +6286,7 @@
       "type": "ANDROID",
       "version": {
         "major": 1,
-        "minor": 0
+        "minor": 1
       },
       "next": null,
       "api": [
@@ -6299,6 +6299,18 @@
         {
           "name": "godot_android_get_activity",
           "return_type": "jobject",
+          "arguments": [
+          ]
+        },
+        {
+          "name": "godot_android_get_surface",
+          "return_type": "jobject",
+          "arguments": [
+          ]
+        },
+        {
+          "name": "godot_android_is_activity_resumed",
+          "return_type": "bool",
           "arguments": [
           ]
         }
