+++ modules/gdnative/include/android/godot_android.h
@@ -46,6 +46,8 @@ extern "C" {
 
 JNIEnv *GDAPI godot_android_get_env();
 jobject GDAPI godot_android_get_activity();
+jobject GDAPI godot_android_get_surface();
+bool GDAPI godot_android_is_activity_resumed();
 
 #ifdef __cplusplus
 }
