+++ platform/android/java/src/org/godotengine/godot/GodotRenderer.java
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  android_gdn.cpp                                                      */
+/*  GodotRenderer.java                                                   */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,59 +28,34 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "modules/gdnative/gdnative.h"
+package org.godotengine.godot;
 
-// Code by Paritosh97 with minor tweaks by Mux213
-// These entry points are only for the android platform and are simple stubs in all others.
+import android.opengl.GLSurfaceView;
+import javax.microedition.khronos.egl.EGLConfig;
+import javax.microedition.khronos.opengles.GL10;
+import org.godotengine.godot.utils.GLUtils;
 
-#ifdef __ANDROID__
-#include "platform/android/java_godot_wrapper.h"
-#include "platform/android/os_android.h"
-#include "platform/android/thread_jandroid.h"
-#else
-#define JNIEnv void
-#define jobject void *
-#endif
+/**
+ * Godot's renderer implementation.
+ */
+class GodotRenderer implements GLSurfaceView.Renderer {
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+	public void onDrawFrame(GL10 gl) {
+		GodotLib.step();
+		for (int i = 0; i < Godot.singleton_count; i++) {
+			Godot.singletons[i].onGLDrawFrame(gl);
+		}
+	}
 
-JNIEnv *GDAPI godot_android_get_env() {
-#ifdef __ANDROID__
-	return ThreadAndroid::get_env();
-#else
-	return NULL;
-#endif
+	public void onSurfaceChanged(GL10 gl, int width, int height) {
+
+		GodotLib.resize(width, height);
+		for (int i = 0; i < Godot.singleton_count; i++) {
+			Godot.singletons[i].onGLSurfaceChanged(gl, width, height);
+		}
+	}
+
+	public void onSurfaceCreated(GL10 gl, EGLConfig config) {
+		GodotLib.newcontext(GLUtils.use_32);
+	}
 }
-
-jobject GDAPI godot_android_get_activity() {
-#ifdef __ANDROID__
-	OS_Android *os_android = (OS_Android *)OS::get_singleton();
-	return os_android->get_godot_java()->get_activity();
-#else
-	return NULL;
-#endif
-}
-
-jobject GDAPI godot_android_get_surface() {
-#ifdef __ANDROID__
-	OS_Android *os_android = (OS_Android *)OS::get_singleton();
-	return os_android->get_godot_java()->get_surface();
-#else
-	return NULL;
-#endif
-}
-
-bool GDAPI godot_android_is_activity_resumed() {
-#ifdef __ANDROID__
-	OS_Android *os_android = (OS_Android *)OS::get_singleton();
-	return os_android->get_godot_java()->is_activity_resumed();
-#else
-	return false;
-#endif
-}
-
-#ifdef __cplusplus
-}
-#endif
